// src/enums/common.ts
export enum UPLOAD_FILE_TYPE {
  AUDIO = 'Audio',
  GALLERY = 'Gallery',
  CAMERA = 'Camera',
  UPLOAD_DOCUMENT = 'Upload Document',
  CALENDAR_EVENT = 'Calendar Event',
  TASK = 'Task',
  NOTE = 'Note',
  EMAIL = 'Email',
}

export enum CHAT_MESSAGE_TYPE {
  TEXT = 'TEXT',
  SYSTEM_USER_ADDED = 'SYSTEM_USER_ADDED',
  SYSTEM_USER_REMOVED = 'SYSTEM_USER_REMOVED',
  SYSTEM_USER_LEFT = 'SYSTEM_USER_LEFT',
  SYSTEM_USER_JOINED = 'SYSTEM_USER_JOINED',
  TEXT_WITH_ATTACHMENTS = 'TEXT_WITH_ATTACHMENTS',
}